Endpoints e ajustes de validação de estoque

See merge request projetos-integrados-2/devlapa/devlapa!7

+++ o-pai-o/src/main/java/com/devlapa/o_pai_o/domain/vendas/VendasRequestDTO.java
@@ -1,4 +1,6 @@
 package com.devlapa.o_pai_o.domain.vendas;
 
-public record VendasRequestDTO(Long formasPagamentosId, Double valor_total, StatusVenda statusVenda) {
+import java.math.BigDecimal;
+
+public record VendasRequestDTO(Long formasPagamentosId, BigDecimal valor_total, StatusVenda statusVenda) {
 }
